requirements.txt
streamlit
SpeechRecognition
google-generativeai
pandas
SQLAlchemy
psycopg2-binary
pymysql
cryptography
plotly